refactor(auth): replace WorkspaceService with DomainManagerService
Updated the SignInUpService to use DomainManagerService instead of WorkspaceService for subdomain generation. Moved the subdomain generation logic from WorkspaceService to DomainManagerService, centralizing domain management functions and streamlining workspace setup.
